Move solver fallback to root conftest so notebook examples also use GLPK
The Gurobi license check and STANDARD_SOLVER override lived only in
test/conftest.py, so `pytest --nbval-lax examples/` never picked it up.
On runners without a full Gurobi license the notebook examples failed
with "Model too large for size-limited license".

Moving the fallback to a root-level conftest.py makes pytest discover it
for both `pytest test/` and `pytest examples/` invocations.
